Rearranging test in proper packages
