+++ src/test/java/hrp/commons/testcases/LiveServerTestCase.java
@@ -1,0 +1,15 @@
+package hrp.commons.testcases;
+
+import hrp.HomePlannerApp;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootContextLoader;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+@ContextConfiguration(classes = HomePlannerApp.class, loader = SpringBootContextLoader.class)
+public abstract class LiveServerTestCase {
+
+}
